requirements.txt
autoflake==2.0.1
black==23.1.0
deepface==0.0.93
gradio~=5.33.1
grad-cam~=1.5.5
hf-xet~=1.1.3
jupyter==1.1.1
notebook~=7.4.3
pillow~=11.2.1
pre-commit==4.2.0
torch~=2.7.1
transformers~=4.52.4
opencv-python~=4.11.0.86
torchvision~=0.22.1
wandb==0.20.1
scikit-learn==1.7.0
seaborn==0.13.2
lightning==2.5.1.post0